Consolidate insights route
