+++ server.js
@@ -241,56 +241,6 @@ app.delete('/api/goals/:id', authMiddleware, async (req, res) => {
   }
 });
 
-// ✅ Protected: Insights
-app.get('/api/insights', authMiddleware, async (req, res) => {
-  try {
-    const userId = req.user._id;
-    const start = startOfCurrentMonth();
-    const end = endOfCurrentMonth();
-    const topCatAgg = await Expense.aggregate([
-      { $match: { userId, date: { $gte: start, $lte: end } } },
-      { $group: { _id: '$category', total: { $sum: '$amount' } } },
-      { $sort: { total: -1 } },
-      { $limit: 1 }
-    ]);
-    const topCategory = topCatAgg[0]?._id || null;
-
-    const vendorAgg = await Transaction.aggregate([
-      { $match: { userId } },
-      { $group: { _id: '$vendor', count: { $sum: 1 } } },
-      { $sort: { count: -1 } },
-      { $limit: 1 }
-    ]);
-    const frequentVendor = vendorAgg[0]?._id || null;
-
-    const trend = [];
-    const now = new Date();
-    for (let i = 0; i < 3; i++) {
-      const y = now.getFullYear();
-      const m = now.getMonth() - i;
-      const s = startOfMonth(y, m);
-      const e = endOfMonth(y, m);
-      const [inc, exp] = await Promise.all([
-        Income.aggregate([
-          { $match: { userId, date: { $gte: s, $lte: e } } },
-          { $group: { _id: null, total: { $sum: '$amount' } } }
-        ]),
-        Expense.aggregate([
-          { $match: { userId, date: { $gte: s, $lte: e } } },
-          { $group: { _id: null, total: { $sum: '$amount' } } }
-        ])
-      ]);
-      const incTotal = inc[0]?.total || 0;
-      const expTotal = exp[0]?.total || 0;
-      trend.unshift({ month: `${s.getFullYear()}-${s.getMonth() + 1}`, savings: incTotal - expTotal });
-    }
-
-    res.json({ topCategory, frequentVendor, savingsTrend: trend });
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: 'Failed to get insights' });
-  }
-});
 
 const PORT = process.env.PORT || 3000;
 app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
